refactor(firecrawl): consolidate executable search providers
