+++ extensions/firecrawl/index.ts
@@ -1,9 +1,11 @@
 // Firecrawl plugin entrypoint registers its OpenClaw integration.
 import { definePluginEntry, type AnyAgentTool } from "openclaw/plugin-sdk/plugin-entry";
 import { createFirecrawlWebFetchProvider } from "./src/firecrawl-fetch-provider.js";
-import { createFirecrawlFreeWebSearchProvider } from "./src/firecrawl-free-search-provider.js";
 import { createFirecrawlScrapeTool } from "./src/firecrawl-scrape-tool.js";
-import { createFirecrawlWebSearchProvider } from "./src/firecrawl-search-provider.js";
+import {
+  createFirecrawlFreeWebSearchProvider,
+  createFirecrawlWebSearchProvider,
+} from "./src/firecrawl-search-provider.js";
 import { createFirecrawlSearchTool } from "./src/firecrawl-search-tool.js";
 
 export default definePluginEntry({
