+++ extensions/firecrawl/web-search-provider.ts
@@ -1,3 +1,5 @@
 // Firecrawl provider module implements model/runtime integration.
-export { createFirecrawlFreeWebSearchProvider } from "./src/firecrawl-free-search-provider.js";
-export { createFirecrawlWebSearchProvider } from "./src/firecrawl-search-provider.js";
+export {
+  createFirecrawlFreeWebSearchProvider,
+  createFirecrawlWebSearchProvider,
+} from "./src/firecrawl-search-provider.js";
